+++ src/app/concert/[id]/_shared/components/concert-info/concert-info.module.scss
@@ -6,7 +6,7 @@
   justify-content: center;
   align-items: center;
   width: 100%;
-  
+
   .background_container {
     width: calc(100% + 48px);
     height: 247px;
@@ -17,9 +17,9 @@
     overflow: hidden; //blur 밖으로 안나오도록 hidden 처리
 
     .background_image {
-    width: 100%;
-    height: 100%;
-    filter: blur(4px);
+      width: 100%;
+      height: 100%;
+      filter: blur(4px);
     }
   }
