+++ src/app/concert/form/[id]/_shared/services/query-key.ts
@@ -1,4 +1,4 @@
-import { CreateConcertFormRequest } from '@/app/concert/form/_shared/services/type';
+import { CreateConcertFormRequest } from '@/app/concert/form/[id]/_shared/services/type';
 
 const queryKey = {
   createConcertForm: (request: CreateConcertFormRequest) => [
