+++ src/app/concert/form/page.tsx
@@ -2,7 +2,7 @@
 
 import AppBarSetter from '@/shared/components/header/app-bar/app-bar-setter';
 
-import Form from './_shared/components/form/form';
+import Form from './[id]/_shared/components/form/form';
 import styles from './page.module.scss';
 
 export default function Page() {
