+++ src/app/history/_shared/components/history-list/history-list.module.scss
@@ -5,13 +5,12 @@
   flex-direction: column;
   justify-content: center;
   width: 100%;
-  
-  .count{
+
+  .count {
     @include title-sm;
     color: var(--textColor-main);
     margin-top: 40px;
   }
-
 
   .asterisk {
     @include title-sm;
