+++ src/app/history/page.module.scss
@@ -5,7 +5,7 @@
   flex-direction: column;
   align-items: flex-start;
 
-  .button_container{
+  .button_container {
     display: flex;
     margin-top: 24px;
     gap: 16px;
