+++ src/shared/components/button/tab-button/tab-button.module.scss
@@ -7,7 +7,7 @@
   align-items: center;
   border-radius: 100px;
   border: 1px solid var(--lineColor-main);
- 
+
   &:active,
   &.active {
     border: 1px solid var(--brandColor-main);
